weekend debugging

+++ Main/helpers.ino
@@ -5,7 +5,7 @@
   This function prints a single into to serial followed by a linebreak
 */
 
-void printIntToSerial(float valueToPrint) {
+void printIntToSerial(int valueToPrint) {
 
   Serial.print(valueToPrint);
   Serial.print("\n");
